+++ srcs/user.cpp
@@ -87,7 +87,7 @@ void user::setNick(std::string nick, std::vector<user> users)
 {
     bool firstSet = false;
     std::string oldNick = this->nick;
-    for (int i = 0; i < users.size(); i++)
+    for (size_t i = 0; i < users.size(); i++)
     {
         if (users[i].getNick() == nick)
         {
